User registration
Set up the Login/register screen on the home.html page.
Modified the scss to allow for hiding the required elements. For some reason the btns wil show up when login or register is clicked but it takes a second click for the input fields to show up.
Users are able to register with the website by providing a email and password. the email and password are html samitzed on the back end.
I messed arround with cors abit. I'm not really sure if the problem is fixed.