1.c and 1.f
I will work on switch everything over from bcrypt to argon2.
Error msg shows up when the user enters a wrong email + password. Error message is the same if the email is not in the database of if the email is in but the password is wrong. if the account is disabled, a message will show up with that warning.