The user is now able to submit songs and Song reviews
The front end sends the validation token through the body.
I was unable to get the front end to send the token from the headers.

The site is now able to make new songs.
All ID3v1 attributes are stored  with the song. Some attributes are taken from input, some are generated at the time of submission.
Title and Artist are enforced fields.
You are also now able to add a review to an existing song.
The review also erquires a rating to be given to the song.  This updates the existing songs stats ( number of reviews, avg rating, total review count) as well.