Setting admin and deactivation
Work was done to allow for an admin/ site manager to set other users accounts to be deactivated or as admin or as both. However, in this instance being both deactivated and admin, deactivation takes priority and you cannot sign in.
Some ui work was done to meet the test case requirements.